Move extra metadata parsers into their own module
