+++ cidc_schemas/prism/__init__.py
@@ -3,10 +3,9 @@ from .merger import (
     merge_artifact,
     merge_artifact_extra_metadata,
     merge_clinical_trial_metadata,
-    parse_elisa,
-    parse_npx,
     InvalidMergeTargetException,
     MergeCollisionException,
 )
+from .extra_metadata import parse_elisa, parse_npx
 from .pipelines import generate_analysis_configs_from_upload_patch
 from .constants import *
