+++ cidc_schemas/prism/merger.py
@@ -10,6 +10,7 @@ from jsonmerge import Merger, strategies
 
 from ..json_validation import load_and_validate_schema
 from ..util import get_path, get_source
+from .extra_metadata import EXTRA_METADATA_PARSERS
 from .constants import PROTOCOL_ID_FIELD_NAME
 
 logger = logging.getLogger(__file__)
@@ -135,9 +136,9 @@ def merge_artifact_extra_metadata(
         additional_artifact_metadata: relevant metadata collected while updating artifact
     """
 
-    if assay_hint not in _EXTRA_METADATA_PARSERS:
+    if assay_hint not in EXTRA_METADATA_PARSERS:
         raise Exception(f"Assay {assay_hint} does not support extra metadata parsing")
-    extract_metadata = _EXTRA_METADATA_PARSERS[assay_hint]
+    extract_metadata = EXTRA_METADATA_PARSERS[assay_hint]
 
     artifact_extra_md_patch = extract_metadata(extra_metadata_file)
     return _update_artifact(ct, artifact_extra_md_patch, artifact_uuid)
@@ -239,111 +240,3 @@ def merge_clinical_trial_metadata(patch: dict, target: dict) -> (dict, List[str]
     merged = merger.merge(target, patch)
 
     return merged, list(validator.iter_errors(merged))
-
-
-# Build a regex from the CIMAC ID pattern in the schema
-cimac_id_regex = re.compile(
-    load_and_validate_schema("sample.json")["properties"]["cimac_id"]["pattern"]
-)
-
-
-def parse_elisa(xlsx: BinaryIO) -> dict:
-    """
-    Parses the given ELISA grand serology results file to extract a list of sample IDs.
-    If the file is not valid NPX but still xlsx the function will
-    return a dict containing an empty list. Sample IDs not conforming to the CIMAC ID
-    format will be skipped. The function will pass along any IO errors.
-    Args:
-        xlsx: an opened NPX file
-    Returns:
-        arg1: a dict of containing list of sample IDs and number of samples
-    """
-
-    # load the file
-    if type(xlsx) == str:
-        raise TypeError(f"parse_npx only accepts BinaryIO and not file paths")
-
-    workbook = openpyxl.load_workbook(xlsx)
-
-    # extract data to python
-    ids = []
-    worksheet = workbook[workbook.sheetnames[0]]
-
-    for i, row in enumerate(worksheet.iter_rows()):
-
-        if i == 0:
-            assert "CIMAC ID" == row[0].value
-            continue
-
-        val = row[0].value
-
-        if val:
-            if cimac_id_regex.match(val):
-                ids.append(val)
-
-    sample_count = len(ids)
-
-    samples = {"samples": ids, "number_of_samples": sample_count}
-
-    return samples
-
-
-def parse_npx(xlsx: BinaryIO) -> dict:
-    """
-    Parses the given NPX file from olink to extract a list of sample IDs.
-    If the file is not valid NPX but still xlsx the function will
-    return a dict containing an empty list. Sample IDs not conforming to the CIMAC ID
-    format will be skipped. The function will pass along any IO errors.
-    Args:
-        xlsx: an opened NPX file
-    Returns:
-        arg1: a dict of containing list of sample IDs and number of samples
-    """
-
-    # load the file
-    if type(xlsx) == str:
-        raise TypeError(f"parse_npx only accepts BinaryIO and not file paths")
-
-    workbook = openpyxl.load_workbook(xlsx)
-
-    # extract data to python
-    ids = []
-    for worksheet_name in workbook.sheetnames:
-
-        # simplify.
-        worksheet = workbook[worksheet_name]
-        seen_onlinkid = False
-        for i, row in enumerate(worksheet.iter_rows()):
-
-            # extract values from row
-            vals = [col.value for col in row]
-
-            first_cell = vals[0]
-
-            # skip empty
-            if len(vals) == 0 or first_cell is None:
-                continue
-
-            # check if we are starting ids
-            if first_cell == "OlinkID":
-                seen_onlinkid = True
-                continue
-
-            # check if we are done.
-            if first_cell == "LOD":
-                break
-
-            # get the identifier
-            if seen_onlinkid:
-                # check that it is a CIMAC ID
-                if cimac_id_regex.match(first_cell):
-                    ids.append(first_cell)
-
-    sample_count = len(ids)
-
-    samples = {"samples": ids, "number_of_samples": sample_count}
-
-    return samples
-
-
-_EXTRA_METADATA_PARSERS = {"olink": parse_npx, "elisa": parse_elisa}
